+++ service/controller/resource/app/desired.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"strconv"
 
+	"github.com/Masterminds/semver"
 	"github.com/ghodss/yaml"
 	g8sv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/microerror"
@@ -58,7 +59,28 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) ([]*g8s
 		userConfig := newUserConfig(cr, appSpec, configMaps, secrets)
 
 		if !appSpec.LegacyOnly {
-			apps = append(apps, r.newApp(*cc, cr, appSpec, userConfig))
+			app := r.newApp(*cc, cr, appSpec, userConfig)
+
+			// For chart-operator only we need to set the Helm major version
+			// label. This is so the correct instance of app-operator processes
+			// the CR.
+			if appSpec.App == chartOperatorAppName {
+				helmMajorVersion := "2"
+
+				version, err := semver.NewVersion(appSpec.Version)
+				if err != nil {
+					return nil, microerror.Mask(err)
+				}
+
+				// For chart-operator 1.0.0 and greater we use Helm 3.
+				if version.Major() >= 1 {
+					helmMajorVersion = "3"
+				}
+
+				app.Labels[label.AppOperatorHelmMajorVersion] = helmMajorVersion
+			}
+
+			apps = append(apps, app)
 		}
 	}